+++ native/comrak_nif/src/types/options.rs
@@ -1,7 +1,7 @@
 mod sanitize;
 
 use autumnus::elixir::ExFormatterOption;
-use comrak::{ExtensionOptions, ListStyleType, ParseOptions, RenderOptions};
+use comrak::options::{Extension, ListStyleType, Parse, Render};
 pub use sanitize::*;
 use std::sync::Arc;
 
@@ -34,9 +34,9 @@ pub struct ExExtensionOptions {
     pub cjk_friendly_emphasis: bool,
 }
 
-impl From<ExExtensionOptions> for ExtensionOptions<'_> {
+impl From<ExExtensionOptions> for Extension<'_> {
     fn from(options: ExExtensionOptions) -> Self {
-        ExtensionOptions {
+        Extension {
             strikethrough: options.strikethrough,
             tagfilter: options.tagfilter,
             table: options.table,
@@ -78,15 +78,19 @@ pub struct ExParseOptions {
     pub default_info_string: Option<String>,
     pub relaxed_tasklist_matching: bool,
     pub relaxed_autolinks: bool,
+    pub ignore_setext: bool,
+    pub tasklist_in_table: bool,
 }
 
-impl From<ExParseOptions> for ParseOptions<'_> {
+impl From<ExParseOptions> for Parse<'_> {
     fn from(options: ExParseOptions) -> Self {
-        ParseOptions {
+        Parse {
             smart: options.smart,
             default_info_string: options.default_info_string,
             relaxed_tasklist_matching: options.relaxed_tasklist_matching,
             relaxed_autolinks: options.relaxed_autolinks,
+            ignore_setext: options.ignore_setext,
+            tasklist_in_table: options.tasklist_in_table,
             broken_link_callback: None,
         }
     }
@@ -116,12 +120,11 @@ pub struct ExRenderOptions {
     pub github_pre_lang: bool,
     pub full_info_string: bool,
     pub width: usize,
-    pub unsafe_: bool,
+    pub r#unsafe: bool,
     pub escape: bool,
     pub list_style: ExListStyleType,
     pub sourcepos: bool,
     pub escaped_char_spans: bool,
-    pub ignore_setext: bool,
     pub ignore_empty_links: bool,
     pub gfm_quirks: bool,
     pub prefer_fenced: bool,
@@ -131,19 +134,18 @@ pub struct ExRenderOptions {
     pub experimental_minimize_commonmark: bool,
 }
 
-impl From<ExRenderOptions> for RenderOptions {
+impl From<ExRenderOptions> for Render {
     fn from(options: ExRenderOptions) -> Self {
-        RenderOptions {
+        Render {
             hardbreaks: options.hardbreaks,
             github_pre_lang: options.github_pre_lang,
             full_info_string: options.full_info_string,
             width: options.width,
-            unsafe_: options.unsafe_,
+            r#unsafe: options.r#unsafe,
             escape: options.escape,
             list_style: ListStyleType::from(options.list_style),
             sourcepos: options.sourcepos,
             escaped_char_spans: options.escaped_char_spans,
-            ignore_setext: options.ignore_setext,
             ignore_empty_links: options.ignore_empty_links,
             gfm_quirks: options.gfm_quirks,
             prefer_fenced: options.prefer_fenced,